Modified alien_invasion.py and ship.py
In the Ship class, added moving_left flag and modified the update() method to move the ship to the left when moving_left flag is True.

In the AlienInvasion class, modified _check_events() method to respond by moving the ship to the left  when the user presses down and releases the left arrow key.