Modified the update() method in the Ship class in ship.py file
Added conditions to check the ship's position before changing the value of self.x and self.y so that the ship does not move pass the screen's edges.

The ship can move right if the value the ship rect's right is less than the value of the screen rect's right.
The ship can move left if the ship rect's left value is greater than the screen rect's left (0).
The ship can move up if the ship rect's top is greater than the screen rect's top (0).
The ship can move down if the ship rect's bottom is less than the screen rect's bottom.