Added fullscreen mode in alien_invasion.py file, change ship's speed in settings.py
Modified the __init__() method to run the game in fullscreen mode. Set the screen surface to origin size (0, 0) and used pygame.FULLSCREEN attribute to figure a window size to fill the screen. Update the setting's screen width and height to the width and heigth value of the screen's rect object.

Change the ship's speed from 1.5 to 3.0 in the Settings class (settings.py).